# backend/api/openapiMAMON.yaml
openapi: 3.0.3
info:
  title: Mamon API
  description: Welcome the the mamon API documentation
  contact:
    email: dev.gam.vollmer@gmail.com
  license:
    name: BSD License
  version: v1
servers:
  - url: http://localhost:8000/api
    description: Local Testing
  - url: https://staging-mamon.esrtheta.nl/api
    description: Testing
  - url: https://mamon.esrtheta.nl/api
    description: Live site
...
paths:
  /login/:
    post:
      description: Log into the site
      requestBody:
        content:
          application/json:
            schema:
              type: object
              required:
                - username
                - password
              properties:
                username:
                  type: string
                password:
                  type: string
      responses:
        "200":
          description: successful operation
          content:
            application/json:
              schema:
                type: object
                properties:
                  user:
                    $ref: "#/components/schemas/UserMe"
                  access_token:
                    type: string
                    description: JWT access token
                  refresh_token:
                    type: string
                    description: JWT refresh token
        "400":
          description: Bad Request
        "401":
          description: Unauthorized
        "403":
          description: Forbidden
      tags:
        - Login
  /users/token/refresh/:
    post:
      summary: get new access token
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                refresh:
                  type: string
                  
                      
      responses:
        "200":
          description: succesful refresh
          content:
            application/json:
              schema:
                type: object
                description: new access token
                properties:
                  access:
                    type: string
                  refresh:
                    type: string

        "422":
          description: invalid refresh token
      parameters: []
      tags:
        - Login
    parameters: []
...
  /cards/{id}:
    get:
      description: ""
      parameters:
        - $ref: "#/components/parameters/id"
      responses:
        "200":
          description: ""
      tags:
        - Cards
    put:
      operationId: api_cards_update
      description: ""
      parameters: []
      responses:
        "200":
          description: ""
      tags:
        - Cards
    delete:
      operationId: api_cards_delete
      description: ""
      parameters: []
      responses:
        "204":
          description: ""
      tags:
        - Cards
    parameters:
      - name: id
        in: path
        required: true
        type: string
...
  /environment/{name}/:
    get:
      operationId: api_environment_read
      description: ""
      parameters: []
      responses:
        "200":
          description: "variable collected"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Environment"

      tags:
        - api
    parameters:
      - name: name
        in: path
        required: true
        type: string
...
  /purchase/{id}:
    get:
      operationId: api_purchase_read
      description: ""
      parameters:
        - $ref: "#/components/parameters/id"
      responses:
        "200":
          description: ""
      tags:
        - Purchases
    parameters:
      - name: id
        in: path
        required: true
        type: string
...
components:
  schemas:
    UserMe:
      allOf:
        - type: object
          properties:
            access:
              type: string
            refresh:
              type: string
...
    Environment:
      type: object
      properties:
        variable: 
          type: string
...
  parameters:
...
    id:
      name: id
      in: path
      required: true
      description: the id of the wanted item
      schema:
        type: integer
...
  securitySchemes:
    Access:
      type: http
      scheme: bearer
      bearerFormat: JWT
    Refresh:
      type: http
      scheme: bearer
...